feat: removed taskStatus under export api
