+++ .gitignore
@@ -110,3 +110,5 @@ dist
 # Jest
 jest_html_reporters.html
 reports
+
+local*.json
